+++ src/Contracts/DataTransferInterface.php
@@ -5,12 +5,12 @@ namespace JuanchoSL\DataTransfer\Contracts;
 interface DataTransferInterface
 {
 
-    public function unset(string $index): self;
+    public function unset(string|int $index): self;
 
-    public function set(string $index, mixed $value): self;
+    public function set(string|int $index, mixed $value): self;
 
-    public function get(string $index, mixed $default = null): mixed;
+    public function get(string|int $index, mixed $default = null): mixed;
 
-    public function has(string $index): bool;
+    public function has(string|int $index): bool;
 
 }