+++ src/Repositories/JsonObjectDataTransfer.php
@@ -8,15 +8,19 @@ use Iterator;
 use JsonSerializable;
 use JuanchoSL\DataTransfer\Contracts\DataTransferInterface;
 
-class JsonObjectDataTransfer extends ObjectDataTransfer implements DataTransferInterface, Iterator, Countable, JsonSerializable
+/**
+ * @implements \Iterator<int|string, DataTransferInterface>
+ */
+class JsonObjectDataTransfer extends JsonArrayDataTransfer implements DataTransferInterface, Iterator, Countable, JsonSerializable
 {
 
     public function __construct(string $json)
     {
-        $body = json_decode($json, false);
+        /*
+        $json = json_decode($json, false);
         if (json_last_error() !== JSON_ERROR_NONE) {
             throw new ExpectationFailedException(json_last_error_msg());
-        }
-        parent::__construct($body);
+        }*/
+        parent::__construct($json);
     }
 }