+++ tests/WriteDataTest.php
@@ -154,9 +154,43 @@ class WriteDataTest extends TestCase
         $this->assertTrue($obj->has('first'), 'Key is setted');
         $this->assertInstanceOf(DataTransferInterface::class, $obj->get('first'), 'Value is instance');
         $this->assertTrue($obj->get('first')->has("index"), 'Subkey is setted');
+        $this->assertEquals("value_0", $obj->first->index->get(0), 'Values are equals');
+        $this->assertEquals("value_0", $obj->get('first')->get('index')->get(0), 'Values are equals');
         foreach ($obj->get('first')->get("index") as $key => $value) {
             $this->assertEquals("value_{$key}", $value, 'Values are equals');
         }
     }
 
+    public function testClonation()
+    {
+        $data = [
+            'index' => ['value_0', 'value_1']
+        ];
+        $obj1 = new JsonObjectDataTransfer(json_encode($data));
+        $obj2 = clone $obj1;
+        $original = json_encode($obj1);
+        $clone = json_encode($obj2);
+        $this->assertEquals($original, $clone);
+        $obj1->index->set(0, 'value_2');
+        $original = json_encode($obj1);
+        $this->assertNotEquals($original, $clone);
+        $this->assertEquals('value_0', $obj2->index->get(0));
+        $this->assertEquals('value_2', $obj1->index->get(0));
+    }
+    public function testClonationObj()
+    {
+        $data = new \stdClass;
+        $data->index = ['value_0', 'value_1'];
+
+        $obj1 = new ObjectDataTransfer($data);
+        $obj2 = clone $obj1;
+        $original = json_encode($obj1);
+        $clone = json_encode($obj2);
+        $this->assertEquals($original, $clone);
+        $obj1->index->set(0, 'value_2');
+        $original = json_encode($obj1);
+        $this->assertNotEquals($original, $clone);
+        $this->assertEquals('value_0', $obj2->index->get(0));
+        $this->assertEquals('value_2', $obj1->index->get(0));
+    }
 }